feat(structured-list): update to v7
