+++ packages/carbon-web-components/src/components/structured-list/structured-list.mdx
@@ -1,4 +1,4 @@
-import { Props, Description } from '@storybook/addon-docs/blocks';
+import { ArgsTable, Description } from '@storybook/addon-docs/blocks';
 import { cdnJs, cdnCss } from '../../globals/internal/storybook-cdn';
 
 # Structured list
@@ -108,7 +108,7 @@ Note: For `boolean` attributes, `true` means simply setting the attribute (e.g.
 `<cds-structured-list has-selection>`) and `false` means not setting the
 attribute (e.g. `<cds-structured-list>` without `has-selection` attribute).
 
-<Props of="cds-structured-list" />
+<ArgsTable of="cds-structured-list" />
 
 ## `<cds-structured-list-row>` attributes and properties
 
@@ -116,4 +116,4 @@ Note: For `boolean` attributes, `true` means simply setting the attribute (e.g.
 `<cds-structured-list-row selected>`) and `false` means not setting the
 attribute (e.g. `<cds-structured-list-row>` without `selected` attribute).
 
-<Props of="cds-structured-list-row" />
+<ArgsTable of="cds-structured-list-row" />
